[7418] Create loadAceEditorAssets util
